Removes `Redirect` as transition option.
While a useful bit of shorthand, I have found `Redirect`ing from one state to another is really buisiness logic and not a part of the design of the state machine itself. So I'm removing it as an option from `shouldTransitionFrom(_, to:)`.

As a result, there's no reason for `shouldTransitionFrom()` not to return a `Bool`. So now it does.