Kills delegates and delegate protocols.
The delegate pattern is so 1989. By ditching the formality around a delegate object (and associated protocol) in favor of a simple closure callback, a lot of the generics-hell cleaned itself up and we're left with a much more straight-forward interface.

Thanks to @algal for pushing me in this direction.

Also adds tests and quite a bit of inline documentation.